Created a user Actor
Created a user Actor class to handle two messages :

1. PlayMovieMessage
2. StopMovieMessage

Goal : The business logic that we are going to implement is that we're going to allow a movie to be played with our existing play movie message and allow a movie to be stopped with our new Stop movie message class.

We are only going to allow a movie to be stopped if there's a movie playing, and we can only start a movie if there isn't one playing in the first place.