Refactored to use switchable behaviours
Our actor has two behaviours it carries out depending on the message that is passed to us. We can either play a movie or stop a movie.

The initial behaviour of our actor is the stopped behaviour, so when we receive a stop playing movie message we get an error, however when we receive a start playing movie message we process that message → so we play that movie, we then change the behaviour of our actor to playing, so that if we receive another start playing movie message we receive an error since the movie is already running.